Simplified the number of files
